# web/swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "API интерфейс SAMPLE",
...
        "contact": {
            "name": "Nikolaj Rudakov",
            "email": "nnrudakov@gmail.com",
            "url": "https://github.com/nnrudakov/sample-api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.sample.nr",
            "description": "Тестовый сервер"
        }
    ],
    "paths": {
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Получение информации о пользователе",
                "description": "Возвращает информацию о пользователе по его идентификатору.\n\nАтрибуты по типам ([см. использование](#section/Poluchenie-atributov-sushnostej)):\n\n| Атрибут | По умолчанию | Расширенные  |\n|---------|:------------:|:------------:|\n| id      | *            |              |\n| email   | *            |              |\n| name    | *            |              |\n| enabled |              | *            |\n| role    |              | *            |",
                "operationId": "users-get-one",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Выбор атрибутов по типам",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Attributes"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Обновление пользователя",
                "description": "Получает новые данные пользователя и обновляет.",
                "operationId": "users-update",
                "parameters": [
                    {
                        "name": "X-Csrf-Token",
                        "in": "header",
                        "description": "CSRF токен",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Новые данные",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/email"
                                    },
                                    "password": {
                                        "title": "Новый пароль",
                                        "description": "Разрешается использовать только буквы английского алфавита, цифры и спец. символы. Минимум 6 символов.",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_repeat": {
                                        "title": "Повторение нового пароля",
                                        "description": "Разрешается использовать только буквы английского алфавита, цифры и спец. символы. Минимум 6 символов.",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "name": {
                                        "$ref": "#/components/schemas/name"
                                    },
                                    "enabled": {
                                        "title": "Активен",
                                        "description": "Игнорируется если происходит обновление главного администратора.",
                                        "type": "boolean"
                                    },
                                    "role": {
                                        "title": "Роль",
                                        "description": "Игнорируется если происходит обновление главного администратора.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Данные успешно изменены"
                    },
                    "400": {
                        "description": "Неверный CSRF токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибки валидации данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidateError"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "Ошибка",
                "properties": {
                    "name": {
                        "title": "Имя ошибки",
                        "description": "Имя ошибки может использоваться в заголовках диалоговых окон.",
                        "type": "string"
                    },
                    "message": {
                        "title": "Текст ошибки",
                        "description": "Текст ошибки для пользователя.",
                        "type": "string"
                    },
                    "code": {
                        "title": "Код ошибки",
                        "description": "Код ошибки внутри системы.<br>0 — Неизвестная ошибка;<br>1 — Объект не найден;<br>2 — Неверный CSRF токен;<br>3 — Неверные параметры запроса;<br>4 — Ошибка входа/выхода в систему;<br>5 — Доступ запрещён",
                        "type": "integer",
                        "format": "int32",
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5
                        ]
                    },
                    "status": {
                        "title": "HTTP статус",
                        "description": "[HTTP](https://en.wikipedia.org/wiki/List_of_HTTP_status_codes) статус ошибки.",
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "title": "Исключение сервера",
                        "description": "Исключение сервера, которое вызвало ошибку. Присутствует только в режиме разработки.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidateError": {
                "title": "Ошибки валидации",
                "properties": {
                    "field": {
                        "title": "Имя поля",
                        "description": "Имя поля, в котором произошла ошибка.",
                        "type": "string"
                    },
                    "message": {
                        "title": "Текст ошибки",
                        "description": "Текст ошибки для пользователя.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Attributes": {
                "title": "Выбор атрибутов по типам",
                "properties": {
                    "fields": {
                        "title": "Атрибуты по умолчанию",
                        "description": "Список атрибутов через запятую.",
                        "type": "string"
                    },
                    "expand": {
                        "title": "Расширенные атрибуты",
                        "description": "Список атрибутов через запятую.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "id": {
                "title": "Идентификатор",
                "type": "integer",
                "format": "int64"
            },
...
            "role": {
                "title": "Роль",
                "description": "Возможные роли:<br>`admin` — Администратор;<br>`user` — Пользователь.",
                "type": "string",
                "enum": [
                    "admin",
                    "user"
                ]
            },
...
            "email": {
                "title": "E-mail",
                "type": "string"
            },
            "name": {
                "title": "Имя",
                "type": "string"
            },
            "User": {
                "title": "Пользователь",
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/id"
                    },
                    "email": {
                        "$ref": "#/components/schemas/email"
                    },
                    "name": {
                        "title": "Имя",
                        "type": "string"
                    },
                    "enabled": {
                        "title": "Активность",
                        "description": "Является ли пользователь не заблокированным.",
                        "type": "boolean"
                    },
                    "role": {
                        "$ref": "#/components/schemas/role"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}